requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3


greenlet==3.2.4

h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10

pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-telegram-bot==22.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3


greenlet==3.2.4
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10

pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-telegram-bot==22.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
sniffio==1.3.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
